feat: [CDS-2993]: k8s release history interface integration (#37776)
* [CDS-2993]: request handlers integrated with release history interface

* [CDS-2993]: some bugfixes and fixed unit tests

* [CDS-2993]: fixed some unit tests and bugs

* [CDS-2993]: k8s delete step integration, with unit test fixes and some bugfixes

* [CDS-2993]: unit test fixes

* [CDS-2993]: use legacy release history as fallback

* [CDS-2993]: some refactors and unit tests

* [CDS-2993]: add created legacy release to legacy release history and integration for k8s traffic split task

* [CDS-2993]: k8s rolling - use imperative release entry for pruning if declarative release history is empty

* [CDS-2993]: integration for CG k8s traffic split and some test fixes

* fix: [CDS-2993]: handle runtimeException thrown by Kubernetes Watch when steady state check is aborted

* fix: [CDS-2993]: K8s Rolling Rollback - FF based logic for getting pod list

* feat: [CDS-2993]: add color label to release secret in BG deployments

* feat: [CDS-2993]: add color label to release secret in BG deployments (CG)

* fix: [CDS-2993]: imperative release history: only add release to release history during rolling deploy if not in canary WF

* fix: [CDS-2993]: add default infradef namespace to list of namespaces and filter null values during pod fetch

* fix: [CDS-2993]: [CDS-47232]: graceful handling of aborted tasks, NPE fix and fix for k8s secret conflict

* Fix imports

* code format

* fix: [CDS-2993]: fix unit tests

* fix: [CDS-2993]: remove redundant release history status update call, remove redundant pruned resource recreation call and continue release numbers from old release history

* fix: [CDS-2993]: some refactors and fixes

* fix: [CDS-2993]: fix unit test

* fix: [CDS-2993]: fix checkstyle

* fix: [CDS-2993]: add missing change for CG

* fix: [CDS-2993]: fail and save current release even when all stages are skipped, modify harness release secret prefix key

* fix: [CDS-2993]: set namespace in resources before pruning diff computation to avoid false positives between old/new release history data

* feat: [CDS-2993]: fix failing tests

* fix: [CDS-2993]: set resources in release only after setting all labels/selectors for all deployment strategies (rolling, BG, canary)

* fix: [CDS-2993]: distinguish b/w empty declarative history cases(failed unsaved first declarative release vs postprod rollback after turning FF on)

* fix: [CDS-2993]: pass FF value in CDNG K8s Delete Step manager request

* fix: [CDS-2993]: fix checkstyle and test compilation error

+++ 930-delegate-tasks/src/main/java/io/harness/helpers/k8s/releasehistory/K8sReleaseHandlerImpl.java
@@ -8,6 +8,7 @@
 package io.harness.helpers.k8s.releasehistory;
 
 import static io.harness.annotations.dev.HarnessTeam.CDP;
+import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
 import static io.harness.k8s.releasehistory.K8sReleaseConstants.RELEASE_KEY;
 import static io.harness.k8s.releasehistory.K8sReleaseConstants.RELEASE_NUMBER_LABEL_KEY;
 import static io.harness.k8s.releasehistory.K8sReleaseConstants.RELEASE_SECRET_TYPE_MAP;
@@ -18,13 +19,17 @@ import static java.lang.String.format;
 import static java.lang.String.join;
 
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.exception.InvalidRequestException;
 import io.harness.k8s.KubernetesContainerService;
+import io.harness.k8s.model.Kind;
 import io.harness.k8s.model.KubernetesConfig;
+import io.harness.k8s.model.KubernetesResourceId;
 import io.harness.k8s.releasehistory.IK8sRelease;
 import io.harness.k8s.releasehistory.IK8sReleaseHistory;
+import io.harness.k8s.releasehistory.K8sBGReleaseHistoryCleanupDTO;
 import io.harness.k8s.releasehistory.K8sRelease;
-import io.harness.k8s.releasehistory.K8sReleaseCleanupDTO;
 import io.harness.k8s.releasehistory.K8sReleaseHistory;
+import io.harness.k8s.releasehistory.K8sReleaseHistoryCleanupDTO;
 import io.harness.k8s.releasehistory.K8sReleaseHistoryHelper;
 import io.harness.k8s.releasehistory.K8sReleasePersistDTO;
 import io.harness.k8s.releasehistory.K8sReleaseSecretHelper;
@@ -35,6 +40,8 @@ import com.google.inject.Singleton;
 import io.kubernetes.client.openapi.models.V1ObjectMetaBuilder;
 import io.kubernetes.client.openapi.models.V1Secret;
 import io.kubernetes.client.openapi.models.V1SecretBuilder;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -77,22 +84,65 @@ public class K8sReleaseHandlerImpl implements K8sReleaseHandler {
   }
 
   @Override
-  public void saveRelease(K8sReleasePersistDTO releasePersistDTO) throws Exception {
-    K8sRelease releaseSecret = (K8sRelease) releasePersistDTO.getRelease();
-    kubernetesContainerService.createOrReplaceSecret(
-        releasePersistDTO.getKubernetesConfig(), releaseSecret.getReleaseSecret());
+  public void saveRelease(K8sReleasePersistDTO releasePersistDTO) throws InvalidRequestException {
+    K8sRelease release = (K8sRelease) releasePersistDTO.getRelease();
+    V1Secret releaseSecret = K8sReleaseSecretHelper.resetSecretVersionMetadata(release.getReleaseSecret());
+
+    kubernetesContainerService.createOrReplaceSecret(releasePersistDTO.getKubernetesConfig(), releaseSecret);
   }
 
   @Override
-  public void cleanReleaseHistory(K8sReleaseCleanupDTO releaseCleanupDTO) throws Exception {
+  public void cleanReleaseHistory(K8sReleaseHistoryCleanupDTO releaseCleanupDTO) throws InvalidRequestException {
     K8sReleaseHistory releaseHistory = (K8sReleaseHistory) releaseCleanupDTO.getReleaseHistory();
     LogCallback logCallback = releaseCleanupDTO.getLogCallback();
     Set<String> releasesToDelete =
         K8sReleaseHistoryHelper.getReleaseNumbersToClean(releaseHistory, releaseCleanupDTO.getCurrentReleaseNumber());
-    logCallback.saveExecutionLog(
-        format("Release numbers to be deleted are: %s", join(SECRET_LABEL_DELIMITER, releasesToDelete)));
+    if (isNotEmpty(releasesToDelete)) {
+      logCallback.saveExecutionLog(
+          format("Release numbers to be deleted are: %s", join(SECRET_LABEL_DELIMITER, releasesToDelete)));
+      deleteReleases(releaseCleanupDTO.getKubernetesConfig(), releaseCleanupDTO.getReleaseName(), releasesToDelete);
+    }
+  }
 
-    deleteReleases(releaseCleanupDTO.getKubernetesConfig(), releaseCleanupDTO.getReleaseName(), releasesToDelete);
+  @Override
+  public void cleanReleaseHistoryBG(K8sBGReleaseHistoryCleanupDTO releaseHistoryCleanupDTO)
+      throws InvalidRequestException {
+    LogCallback logCallback = releaseHistoryCleanupDTO.getLogCallback();
+    List<K8sRelease> releasesToClean = releaseHistoryCleanupDTO.getReleasesToClean()
+                                           .stream()
+                                           .map(K8sRelease.class ::cast)
+                                           .collect(Collectors.toList());
+    if (isNotEmpty(releasesToClean)) {
+      Set<String> releaseNumbersToClean = K8sReleaseHistoryHelper.getReleaseNumbersFromReleases(releasesToClean);
+      logCallback.saveExecutionLog(
+          format("Release numbers to be deleted are: %s", join(SECRET_LABEL_DELIMITER, releaseNumbersToClean)));
+
+      deleteReleases(releaseHistoryCleanupDTO.getKubernetesConfig(), releaseHistoryCleanupDTO.getReleaseName(),
+          releaseNumbersToClean);
+    }
+  }
+
+  @Override
+  public List<KubernetesResourceId> getResourceIdsToDelete(
+      String releaseName, KubernetesConfig kubernetesConfig, LogCallback logCallback) {
+    K8sReleaseHistory releaseHistory = (K8sReleaseHistory) getReleaseHistory(kubernetesConfig, releaseName);
+
+    Map<String, KubernetesResourceId> kubernetesResourceIdMap = new HashMap<>();
+    for (K8sRelease release : releaseHistory.getReleaseHistory()) {
+      release.getResourceIds().forEach(
+          resourceId -> kubernetesResourceIdMap.put(resourceId.namespaceKindNameRef(), resourceId));
+
+      // Create entry for release secret and add to map
+      String releaseSecretName = K8sReleaseSecretHelper.getSecretName(release.getReleaseSecret());
+      KubernetesResourceId releaseSecretResource = KubernetesResourceId.builder()
+                                                       .name(releaseSecretName)
+                                                       .kind(Kind.Secret.name())
+                                                       .namespace(kubernetesConfig.getNamespace())
+                                                       .build();
+      kubernetesResourceIdMap.put(releaseSecretResource.namespaceKindNameRef(), releaseSecretResource);
+    }
+
+    return new ArrayList<>(kubernetesResourceIdMap.values());
   }
 
   private List<K8sRelease> createReleasesFromSecrets(List<V1Secret> releaseSecrets) {
@@ -101,7 +151,8 @@ public class K8sReleaseHandlerImpl implements K8sReleaseHandler {
         .collect(Collectors.toList());
   }
 
-  private void deleteReleases(KubernetesConfig kubernetesConfig, String releaseName, Set<String> releaseNumbers) {
+  private void deleteReleases(KubernetesConfig kubernetesConfig, String releaseName, Set<String> releaseNumbers)
+      throws InvalidRequestException {
     String labelArg = join(SECRET_LABEL_DELIMITER, K8sReleaseSecretHelper.createListBasedArg(RELEASE_KEY, releaseName),
         K8sReleaseSecretHelper.createSetBasedArg(RELEASE_NUMBER_LABEL_KEY, releaseNumbers));
     String fieldArg = K8sReleaseSecretHelper.createCommaSeparatedKeyValueList(RELEASE_SECRET_TYPE_MAP);
